% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPRD_testFramework.R.R
\name{generateInsertSql}
\alias{generateInsertSql}
\title{Get insert SQL}
\usage{
generateInsertSql(databaseSchema = NULL)
}
\value{
list of sql statements
}
\description{
Get insert SQL
}
\details{
Returns insert Sql for native DB.
}
