% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPRD_testFramework.R.R
\name{generateTestSql}
\alias{generateTestSql}
\title{Get testing SQL}
\usage{
generateTestSql(databaseSchema = NULL)
}
\value{
list of sql statements
}
\description{
Get testing SQL
}
\details{
Returns testing Sql for the CDM DB.
}
